# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Orion Belt API
  description: |
    REST and WebSocket API for the Orion Belt privileged access gateway.

    **Base URL:** `/api/v1` (plus `/health` and optional `/metrics` outside the versioned prefix).

    **Authentication** (protected routes — try in order):
    1. `X-API-Key` header
    2. `X-Session-Token` header, `session_token` cookie, or `token` / `access_token` query (WebSocket)
    3. `Authorization: Bearer <JWT>`

    Pubkey login (`/public/login*`) requires a prior `POST /public/auth/challenge` and a signature proving possession of the private key.
    Password login: `POST /public/login/password` (password + TOTP). Console also supports WebAuthn and `osh login` browser bootstrap.

    Optional **SSH CA** endpoints issue short-lived User certs and expose Host CA trust material (`docs/SSH_CA.md`).

    **Admin routes** (`/api/v1/admin/*`) additionally require role `admin` or `operator` (or `is_admin`).

    **Rate limit:** protected (non-admin) routes — default 600 req/min per user (`auth.rate_limit_per_minute`).
  version: "1.0.0"
  contact:
    name: Orion Belt Maintainers
    url: https://github.com/orion-belt-dev/orion-belt
  license:
    name: Apache-2.0 WITH Commons-Clause
    url: https://github.com/orion-belt-dev/orion-belt/blob/master/LICENSE
servers:
  - url: http://localhost:8080
    description: Local / lab default
  - url: "{scheme}://{host}"
    description: Deployed gateway
    variables:
      scheme:
...
      host:
...
paths:
...
  /api/v1/sessions/{id}/watch:
    get:
      tags: [Sessions]
      summary: Live watch an active session (WebSocket)
      description: |
        Upgrades to a WebSocket that streams raw PTY output for an **active**
        session recorded on this gateway node. Auth via `?token=` session or
        `?access_token=` JWT (same as terminal WS). Privileged roles or the
        session owner may watch; connection is read-only.
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "101":
          description: Switching Protocols
        "409":
          description: Session not active or recording not on this node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/terminal/ws:
    get:
      tags: [Terminal]
      summary: Web terminal WebSocket
      description: |
        Upgrade to WebSocket after auth + ReBAC check.
        Creates an audited session with `source=web` and records **PTY output**
        as a timed `.cast` file (keystrokes are not recorded separately).

        On client disconnect the agent channel is closed so the session is marked
        completed and the recording is flushed.

        **Client → server:** binary/text PTY input; JSON resize `{"type":"resize","cols":N,"rows":N}`
        **Server → client:** binary PTY output; text `error: …` on failure
      parameters:
        - name: machine
          in: query
          required: true
          schema: { type: string }
          description: Machine name
        - name: user
          in: query
          schema: { type: string, default: root }
          description: Remote OS user
        - name: token
          in: query
          schema: { type: string }
          description: Session token (alternative to cookie/header)
        - name: access_token
          in: query
          schema: { type: string }
      responses:
        "101":
          description: Switching Protocols (WebSocket)
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: Terminal or recorder unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/files/list:
    get:
      tags: [Files]
      summary: List remote directory
      parameters:
        - $ref: "#/components/parameters/MachineQuery"
        - name: path
          in: query
          schema: { type: string, default: / }
        - $ref: "#/components/parameters/RemoteUserQuery"
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
  /api/v1/files/download:
    get:
      tags: [Files]
      summary: Download remote file
      parameters:
        - $ref: "#/components/parameters/MachineQuery"
        - name: path
          in: query
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RemoteUserQuery"
      responses:
        "200":
          description: File bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /api/v1/files/upload:
    post:
      tags: [Files]
      summary: Upload file to remote path
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [machine, path, file]
              properties:
                machine: { type: string }
                path: { type: string }
                user: { type: string, default: root }
                file: { type: string, format: binary }
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  path: { type: string }
                  size: { type: integer }
  /api/v1/files/mkdir:
    post:
      tags: [Files]
      summary: Create remote directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [machine, path]
              properties:
                machine: { type: string }
                path: { type: string }
                user: { type: string }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/v1/files:
    delete:
      tags: [Files]
      summary: Delete remote path
      parameters:
        - $ref: "#/components/parameters/MachineQuery"
        - name: path
          in: query
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RemoteUserQuery"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/v1/ssh-keys:
    get:
      tags: [SSH Keys]
      summary: List current user SSH public keys
      responses:
        "200":
          description: Keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SSHKey"
    post:
      tags: [SSH Keys]
      summary: Add SSH public key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, public_key]
              properties:
                name: { type: string }
                public_key: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSHKey"
  /api/v1/ssh-keys/{id}:
    delete:
      tags: [SSH Keys]
      summary: Delete SSH key
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SessionHeader:
      type: apiKey
      in: header
      name: X-Session-Token
    SessionCookie:
      type: apiKey
      in: cookie
      name: session_token
    SessionQuery:
      type: apiKey
      in: query
      name: token
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    MachineQuery:
      name: machine
      in: query
      required: true
      schema: { type: string }
      description: Machine name
    RemoteUserQuery:
      name: user
      in: query
      schema: { type: string, default: root }
  responses:
...
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
    Message:
      type: object
      properties:
        message: { type: string }
...
    SSHKey:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        name: { type: string }
        public_key: { type: string }
        key_type: { type: string }
        created_at: { type: string, format: date-time }
    FileListResponse:
      type: object
      properties:
        path: { type: string }
        entries:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              path: { type: string }
              is_dir: { type: boolean }
              size: { type: integer }
              mtime: { type: integer }
        raw: { type: string, description: Fallback listing text }
        error: { type: string }
...
security:
  - ApiKeyAuth: []
  - SessionHeader: []
  - SessionCookie: []
  - BearerJWT: []